+++ src/components/index.ts
@@ -4,7 +4,8 @@ import { PersonCard } from "./PersonCard";
 import { ListCard } from "./ListCard";
 import { SeriesCard } from "./SeriesCard";
 import { SearchBar } from "./SearchBar";
+import { GlobalSearch } from "./GlobalSearch";
 import { Breadcrumbs } from "./Breadcrumbs";
 import { BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState } from "./Skeletons";
 
-export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };
+export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, GlobalSearch, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };